pyproject.toml
[tool.poetry]
name = "element"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
discord-interactive = "^4.0"
pyfiglet = "^0.8.post1"
PyNaCl = "^1.4.0"
python-dotenv = "^0.15.0"
jishaku = "^1.20.0"
discord-musicbot = "^0.0.3-alpha.20170604"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
